Move the `intrinsic_factory` module into the `federated_context` package.
PiperOrigin-RevId: 357995237
