+++ tensorflow_federated/python/core/impl/federated_context/intrinsic_factory_test.py
@@ -19,9 +19,9 @@ import tensorflow as tf
 from tensorflow_federated.python.core.api import computation_types
 from tensorflow_federated.python.core.api import computations
 from tensorflow_federated.python.core.api import intrinsics
-from tensorflow_federated.python.core.impl import intrinsic_factory
 from tensorflow_federated.python.core.impl.context_stack import context_stack_impl
 from tensorflow_federated.python.core.impl.federated_context import federated_computation_context
+from tensorflow_federated.python.core.impl.federated_context import intrinsic_factory
 from tensorflow_federated.python.core.impl.types import placement_literals
